Fixed build warnings
The metadataBase is essentially the root URL of your website and is important for resolving the paths for social open graph or Twitter images correctly, especially in a production environment. It helps search engines understand the correct path for your metadata assets, improving the presentation of your content in search results.